+++ src/win32/cdwin.c
@@ -2343,9 +2343,9 @@ static void set_rotate_attrib(cdCtxCanvas* ctxcanvas, char* data)
 
   if (data)
   {
-    sscanf(data, "%g %d %d", &ctxcanvas->rotate_angle,
-                             &ctxcanvas->rotate_center_x,
-                             &ctxcanvas->rotate_center_y);
+    sscanf(data, "%lg %d %d", &ctxcanvas->rotate_angle,
+                              &ctxcanvas->rotate_center_x,
+                              &ctxcanvas->rotate_center_y);
   }
   else
   {
@@ -2364,7 +2364,7 @@ static char* get_rotate_attrib(cdCtxCanvas* ctxcanvas)
   if (!ctxcanvas->rotate_angle)
     return NULL;
 
-  sprintf(data, "%g %d %d", (double)ctxcanvas->rotate_angle,
+  sprintf(data, "%g %d %d", ctxcanvas->rotate_angle,
                             ctxcanvas->rotate_center_x,
                             ctxcanvas->rotate_center_y);
